src/backend/requirements.txt
# run in the terminal to download following dependencies: pip install -r src/backend/requirements.txt 
fastapi
uvicorn
python-multipart
pillow
openai
dotenv
